fix bug for author to anly delete the posts they created

+++ lib/spa_web/live/article_live/index.ex
@@ -43,6 +43,14 @@ defmodule SpaWeb.ArticleLive.Index do
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
     article = Articles.get_article!(id)
+    current_user_id = socket.assigns.current_user.id
+
+    if article.author_id != current_user_id do
+      raise """
+      An unauthorized user #{current_user_id} tried to delete article #{article.id}
+      """
+    end
+
     {:ok, _} = Articles.delete_article(article)
 
     {:noreply, stream_delete(socket, :articles, article)}
